+++ CHANGES.rst
@@ -12,6 +12,7 @@ Version 2.4.2, 2015-09-16
 =========================
 
 - Fix version conflicts due to too tight pinning, issue #69
+- Added small template for unittest
 
 Version 2.4.1, 2015-09-09
 =========================